LLaMA3.1_exp/requirements.txt
yarl==1.9.4
xxhash==3.5.0
wrapt==1.16.0
tzdata==2024.1
typing-extensions==4.12.2
triton==3.0.0
transformers==4.44.1
tqdm==4.66.5
torch==2.4.0
tokenizers==0.19.1
sympy==1.13.2
stop-words==2018.7.23
soupsieve==2.6
smart-open==7.0.4
scipy==1.10.1
safetensors==0.4.4
requests==2.32.3
regex==2024.7.24
pyyaml==6.0.2
pytz==2024.1
python-dateutil==2.9.0.post0
pyspellchecker==0.8.1
pyarrow==17.0.0
psutil==6.0.0
pandas==2.0.3
packaging==24.1
nvidia-nvtx-cu12==12.1.105
nvidia-nvjitlink-cu12==12.6.20
nvidia-nccl-cu12==2.20.5
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-curand-cu12==10.3.2.106
nvidia-cufft-cu12==11.0.2.54
nvidia-cudnn-cu12==9.1.0.70
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cublas-cu12==12.1.3.1
numpy==1.24.4
nltk==3.9.1
networkx==3.1
multiprocess==0.70.16
multidict==6.0.5
mpmath==1.3.0
markupsafe==2.1.5
joblib==1.4.2
jinja2==3.1.4
huggingface-hub==0.24.6
gensim==4.3.3
fsspec==2024.6.1
frozenlist==1.4.1
filelock==3.15.4
faiss-gpu==1.7.2
dill==0.3.8
deep-translator==1.11.4
datasets==2.21.0
click==8.1.7
charset-normalizer==3.3.2
beautifulsoup4==4.12.3
attrs==24.2.0
async-timeout==4.0.3
aiosignal==1.3.1
aiohttp==3.10.5
aiohappyeyeballs==2.4.0
accelerate==0.33.0
wheel==0.34.2
urllib3==1.25.8
ufw==0.36
ssh-import-id==5.10
six==1.14.0
setuptools==45.2.0
pygobject==3.36.0
pip==20.0.2
idna==2.8
distro==1.4.0
dbus-python==1.2.16
chardet==3.0.4
certifi==2019.11.28
